Move to async/await, remove Node 6 compatibility
