+++ lib/generator.js
@@ -12,8 +12,8 @@ module.exports = class BaseGenerator extends Generator {
     this.defaultConfig = this.fs.readJSON(defaultConfig, {});
     this.props = opts.props || {};
 
-    if (!semver.satisfies(process.version, '>= 6.0.0')) {
-      this.log.error('The generator will only work with Node v6.0.0 and up!');
+    if (!semver.satisfies(process.version, '>= 8.0.0')) {
+      this.log.error('The generator will only work with Node v8.0.0 and up!');
       process.exit();
     }
 
@@ -35,14 +35,6 @@ module.exports = class BaseGenerator extends Generator {
     }
   }
 
-  get hasAsync() {
-    const asyncNode = '>= 8.0.0';
-    const pkgEngineAsync = !(this.pkg.engines && this.pkg.engines.node) ||
-      semver.intersects(this.pkg.engines.node, asyncNode);
-
-    return semver.satisfies(process.version, asyncNode) && pkgEngineAsync;
-  }
-
   get libDirectory() {
     return this.pkg.directories && this.pkg.directories.lib;
   }
@@ -52,8 +44,7 @@ module.exports = class BaseGenerator extends Generator {
   }
 
   _packagerInstall(deps, options) {
-    const packager = this.pkg.engines && this.pkg.engines.yarn ? 
-      'yarn' : 'npm';
+    const packager = this.pkg.engines && this.pkg.engines.yarn ? 'yarn' : 'npm';
     const method = `${packager}Install`;
     const isDev = options.saveDev;
     const existingDependencies = this.pkg[isDev ? 'devDependencies' : 'dependencies'] || {};
